XWAV header parsing
It would be good to support the XWAV format used in the public data set

https://console.cloud.google.com/marketplace/product/noaa-public/passive_acoustic_monitoring

These files are gigabytes large and collect sometimes noncontiguous
shorter (75s) clips with a custom WAV header that associates file
offsets with UTC times. When extracting even shorter clips appropriate
for training and joining them to labels, we will need the UTC time
reference and for clips to avoid crossing boundaries of the 75s
"subchunks."

This patch adds support for parsing the XWAV-specific information from
WAV headers or FLAC metadata.

Change-Id: I4daaa3483d0bd86373babaf3a803c912fb017073