Adds a non-test XWAV writer
In developing UTC endpoints for examplegen, the real XWAVs, in the GB
range, make test runs impractically long. This will be useful to create
test inputs.

test_xwav.py could also probably be refactored to use this instead of
insert_harp_chunk, but that will be left for a separate change.

Change-Id: If739cbe2950e8c27c4219f62051132ee4bd8f3e9