Spectrogram front end for Keras
The module added in this change wraps the transformation of a raw
waveform to an image input for a convolutional neural network in a Keras
layer. This can enable training end-to-end from the raw waveform to
class probabilities or, alternatively, serve as a module to reliably
apply the same spectrogram settings in preprocessing and inference.

Change-Id: Ic9b51d3c288baf1faddb38ae95762555352dafe9